- renamed Utils - added package info with info data for future developers - added java-dock to some Util classes
