+++ exec/java-exec/src/main/java/org/apache/drill/exec/store/pcap/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * For comments on realization of this format plugin look at :
+ * @see <a href="https://issues.apache.org/jira/browse/DRILL-5432"> Jira</a>
+ * @see <a href="https://github.com/apache/drill/pull/831"> pull request</a>
+ */
+package org.apache.drill.exec.store.pcap;
